Entreprise

+++ projet/apiGateway/src/src/main/java/main/src/dto/EntrepriseDTO.java
@@ -4,13 +4,13 @@ public class EntrepriseDTO {
     private int pkMenu;
     private String nom;
     private String image;
-    private Integer prix_unitaire;
-    private Integer quantite;
-    private Integer fkEntreprise;
+    private float prix_unitaire;
+    private int quantite;
+    private int fkEntreprise;
 
     public EntrepriseDTO() {}
 
-    public EntrepriseDTO(Integer pk_menu, String nom, String image, Integer prix_unitaire, Integer quantite, Integer fkEntreprise) {
+    public EntrepriseDTO(int pk_menu, String nom, String image, float prix_unitaire, int quantite, int fkEntreprise) {
         this.nom = nom;
         this.image = image;
         this.prix_unitaire = prix_unitaire;
@@ -20,11 +20,11 @@ public class EntrepriseDTO {
 
     // Getters & Setters
 
-    public Integer getPkMenu() {
+    public int getPkMenu() {
         return pkMenu;
     }
 
-    public void setPkMenu(Integer pkMenu) {
+    public void setPkMenu(int pkMenu) {
         this.pkMenu = pkMenu;
     }
 
@@ -44,27 +44,27 @@ public class EntrepriseDTO {
         this.image = image;
     }
 
-    public Integer getPrix() {
+    public float getPrix() {
         return prix_unitaire;
     }
 
-    public void setPrix(Integer prix_unitaire) {
+    public void setPrix(float prix_unitaire) {
         this.prix_unitaire = prix_unitaire;
     }
 
-    public Integer getFkEntreprise() {
+    public int getFkEntreprise() {
         return fkEntreprise;
     }
 
-    public void setFkEntreprise(Integer fkEntreprise) {
+    public void setFkEntreprise(int fkEntreprise) {
         this.fkEntreprise = fkEntreprise;
     }
 
-    public Integer getQuantite() {
+    public int getQuantite() {
         return quantite;
     }
 
-    public void setQuantite(Integer quantite) {
+    public void setQuantite(int quantite) {
         this.quantite = quantite;
     }
 }
